Update Program.cs
trying to use double so I can get decimal points

+++ CalcRoom/Program.cs
@@ -6,7 +6,7 @@ namespace CalcRoom // area cost calculation.
     {
         public static void Main()
         {
-            int length, width, area, cost, price;
+            double length, width, area, cost, price;
             Console.Write("Enter length of room in feet: ");
             length = Convert.ToInt32(Console.ReadLine());
             Console.Write("Enter width of room in feet:");
